--- packages/client/src/pages/Home/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './UpperBar';
