+++ packages/client/src/widgets/underBar/UnderBar.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { Button } from 'shared/ui';
-import { Title01 } from '../styles';
+import { Title01 } from '../../shared/ui/styles';
 import PlanetEditIcon from '@icons/icon-planetedit-24-white.svg';
 import PlanetEditIconGray from '@icons/icon-planetedit-24-gray.svg';
 import AddIcon from '@icons/icon-add-24-white.svg';
