+++ packages/client/src/widgets/upperBar/UpperBar.tsx
